requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
youtube-transcript-api==1.2.2
openai==1.3.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6
requests==2.31.0
supabase==2.0.2